requirements-dev.txt
-r requirements.txt
pytest
pytest-html
ruff
psycopg2-binary